pyproject.toml
[tool.poetry]
name = "ewx_pws"
version = "0.1.0"
description = "Support for MSU Enviroweather Network Personal Weather Stations"
authors = ["Patrick Bills"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
python-dotenv = "^0.21.1"
jupyter = "^1.0.0"
notebook = "^6.5.2"
sphinx-autoapi = "^2.0.1"
sphinx-rtd-theme = "^1.2.0"
myst-nb = {version = "^0.17.1", python = "^3.9"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
